+++ apps/backend/src/cards/cards.module.ts
@@ -2,10 +2,10 @@ import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
 import { Card, CardSchema } from 'src/shared/schemas/card.schema';
 
-import { CardsSortingService } from './cards-sorting.service';
 import { CardsController } from './cards.controller';
 import { CardsResolver } from './cards.resolver';
-import { CardsService } from './cards.service';
+import { CardsSortingService } from './services/cards-sorting.service';
+import { CardsService } from './services/cards.service';
 
 @Module({
   imports: [MongooseModule.forFeature([{ name: Card.name, schema: CardSchema }])],
