+++ apps/backend/src/cards/services/cards.service.ts
@@ -3,7 +3,7 @@ import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { Card, CardDocument } from 'src/shared/schemas/card.schema';
 
-import { CreateCardDto } from './dto/create-card.dto';
+import { CreateCardDto } from '../dto/create-card.dto';
 
 @Injectable()
 export class CardsService {
